Separate testing files into their own respective packages, add test cases for SolarSystem and Planet, and add javadocs for Coordinate.
