+++ android/app/src/main/java/edu/gatech/cs2340/rajsjarfiles/spacetrader/entity/universe/Coordinate.java
@@ -1,22 +1,43 @@
 package edu.gatech.cs2340.rajsjarfiles.spacetrader.entity.universe;
 
+/**
+ * Represents a coordinate in the universe.
+ */
 public class Coordinate {
     private int x;
     private int y;
 
+    /**
+     * Two arg constructor for a coordinate.
+     *
+     * @param x the x coordinate
+     * @param y the y coordinate
+     */
     public Coordinate(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
+    /**
+     * @return the x coordinate
+     */
     public int getX() {
         return x;
     }
 
+    /**
+     * @return the y coordinate
+     */
     public int getY() {
         return y;
     }
 
+    /**
+     * Returns distance to another coordinate.
+     *
+     * @param c the other coordinate
+     * @return the distance
+     */
     public double distTo(Coordinate c) {
         int dx = c.x - x;
         int dy = c.y - y;
@@ -43,6 +64,13 @@ public class Coordinate {
         return h;
     }
 
+    /**
+     * Static method for finding the distance between two coordinates.
+     *
+     * @param c1 the first coordinate
+     * @param c2 the second coordinate
+     * @return the distance between the two coordinates
+     */
     public static double distBetween(Coordinate c1, Coordinate c2) {
         int dx = c1.x - c2.x;
         int dy = c1.y - c2.y;
