+++ android/app/src/test/java/edu/gatech/cs2340/rajsjarfiles/spacetrader/entity/player/PlayerTester.java
@@ -1,4 +1,4 @@
-package edu.gatech.cs2340.rajsjarfiles.spacetrader.entity;
+package edu.gatech.cs2340.rajsjarfiles.spacetrader.entity.player;
 
 import org.junit.Test;
